LPS-130022 Extract auxiliary functions
